src/main/java/com/epam/javast/quadrilateral/validator/quadrilateralvalidator/RectangleValidator.java
package com.epam.javast.quadrilateral.validator.quadrilateralvalidator;
import com.epam.javast.quadrilateral.entity.Point;
import com.epam.javast.quadrilateral.entity.Quadrilateral;
import com.epam.javast.quadrilateral.logic.CalculatorHelper;

public class RectangleValidator {
    private CalculatorHelper calculatorHelper = new CalculatorHelper();

    public boolean isValidRectangle(Quadrilateral quadrilateral) {
        if (quadrilateral == null) {
            return false;
        }

        Point pointA = quadrilateral.getPointA();
        Point pointB = quadrilateral.getPointB();
        Point pointC = quadrilateral.getPointC();
        Point pointD = quadrilateral.getPointD();

        double sideAB = calculatorHelper.distance(pointA, pointB);
        double sideBC = calculatorHelper.distance(pointB, pointC);
        double sideCD = calculatorHelper.distance(pointC, pointD);
        double sideDA = calculatorHelper.distance(pointD, pointA);

        double diagonalAC = calculatorHelper.distance(pointA, pointC);
        double diagonalBD = calculatorHelper.distance(pointB, pointD);

        if (sideAB == sideCD && sideBC == sideDA) {
            return diagonalAC == diagonalBD;
        }

        return false;
    }
}
